Multiple changes
Closed the function if(n < 1) with the curly bracket. Changed modulus to = 1. Spaced in the for function between the i = 1. Removed the &&s between the for functions, I don't recall if I added those earlier myself but I removed them. Changed the log10(n) to instead be the log10 of (n^2). That way it takes in all the digits of the squared number not the original integer.